# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.DefaultResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
...
  subject.SubjectRequest:
    properties:
      code:
        type: string
      description:
        type: string
      level_id:
        description: Boleh kosong
        type: string
      major_id:
        description: Boleh kosong (umum)
        type: string
      name:
        type: string
    required:
    - code
    - name
    type: object
  subject.SubjectResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      level_id:
        type: string
      major_id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Documentation for LinguaLoop E-Learning Platform.
  title: LinguaLoop API
  version: "1.0"
paths:
...
  /subjects:
    get:
      description: Get a list of all subjects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/subject.SubjectResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Get all subjects
      tags:
      - Subjects
    post:
      consumes:
      - application/json
      description: Add a new subject to the system
      parameters:
      - description: Subject Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/subject.SubjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/subject.SubjectResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Create new subject
      tags:
      - Subjects
  /subjects/{id}:
    delete:
      description: Soft delete a subject by ID
      parameters:
      - description: Subject ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Delete subject
      tags:
      - Subjects
    get:
      description: Get detailed information of a subject by its ID
      parameters:
      - description: Subject ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/subject.SubjectResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Get subject by ID
      tags:
      - Subjects
    put:
      consumes:
      - application/json
      description: Update existing subject data
      parameters:
      - description: Subject ID
        in: path
        name: id
        required: true
        type: string
      - description: Subject Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/subject.SubjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/subject.SubjectResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Update subject
      tags:
      - Subjects
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"